requirements.txt
blinker==1.9.0
boto3==1.38.38
botocore==1.38.38
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pillow==11.2.1
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
